Define shared stimulus controllers mixin for activities tab
